refactor(rq,sdb,nq): consolidate IncidentCategory in rq
Three Wave 1 task-owners independently defined the five-variant
severity-category enum. Promote rq's `IncidentCategory` to the canonical
type for the status-reporting subsystem and re-export it from sdb (as
`Category`) and nq (as `SeverityCategory`).

- Add `ALL`, `as_str()`, `Display`, `FromStr`, and `InvalidCategory` to
  rq::IncidentCategory so downstream crates do not need to wrap it.
- sdb depends on rq and re-exports the enum; sdb's category tests now
  exercise the canonical type via the local alias.
- nq depends on rq and re-exports the enum; the placeholder
  reconciliation note is removed.

Wire format (SCREAMING_SNAKE_CASE) and severity ordering (Crash is max)
are preserved.
