+++ crates/nq/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2024"
 chrono = { version = "0.4.43", features = ["serde"] }
 futures-util = "0.3.31"
 lapin = "3.7.2"
+rq = { path = "../rq" }
 serde = { version = "1.0.228", features = ["derive"] }
 serde_json = "1.0.145"
 thiserror = "2.0.18"
