+++ crates/sdb/Cargo.toml
@@ -8,6 +8,7 @@ chrono = { version = "0.4.43", features = ["serde"] }
 futures = "0.3.31"
 futures-util = "0.3.31"
 ids = { path = "../ids" }
+rq = { path = "../rq" }
 scylla = { version = "1.5.0", features = ["chrono-04"] }
 serde = { version = "1.0.228", features = ["derive"] }
 thiserror = "2.0.18"
